+++ tests/model_tests/app/forms.py
@@ -1,0 +1,3 @@
+from django.conf import settings
+
+settings.FORMS_IMPORTED=True